create consumer_lifecycle.go with launch/stop logic
