+++ x/ccv/provider/keeper/permissionless.go
@@ -6,8 +6,6 @@ import (
 	"strconv"
 	"time"
 
-	errorsmod "cosmossdk.io/errors"
-	storetypes "cosmossdk.io/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	govv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	"github.com/cosmos/interchain-security/v5/x/ccv/provider/types"
@@ -211,149 +209,6 @@ func (k Keeper) IsConsumerActive(ctx sdk.Context, consumerId string) bool {
 		phase == types.ConsumerPhase_CONSUMER_PHASE_LAUNCHED
 }
 
-// GetConsumerStopTime returns the stop time associated with the to-be-stopped chain with consumer id
-func (k Keeper) GetConsumerStopTime(ctx sdk.Context, consumerId string) (time.Time, error) {
-	store := ctx.KVStore(k.storeKey)
-	buf := store.Get(types.ConsumerIdToStopTimeKey(consumerId))
-	if buf == nil {
-		return time.Time{}, fmt.Errorf("failed to retrieve stop time for consumer id (%s)", consumerId)
-	}
-	var time time.Time
-	if err := time.UnmarshalBinary(buf); err != nil {
-		return time, fmt.Errorf("failed to unmarshal stop time for consumer id (%s): %w", consumerId, err)
-	}
-	return time, nil
-}
-
-// SetConsumerStopTime sets the stop time associated with this consumer id
-func (k Keeper) SetConsumerStopTime(ctx sdk.Context, consumerId string, stopTime time.Time) error {
-	store := ctx.KVStore(k.storeKey)
-	buf, err := stopTime.MarshalBinary()
-	if err != nil {
-		return fmt.Errorf("failed to marshal stop time (%+v) for consumer id (%s): %w", stopTime, consumerId, err)
-	}
-	store.Set(types.ConsumerIdToStopTimeKey(consumerId), buf)
-	return nil
-}
-
-// DeleteConsumerStopTime deletes the stop time associated with this consumer id
-func (k Keeper) DeleteConsumerStopTime(ctx sdk.Context, consumerId string) {
-	store := ctx.KVStore(k.storeKey)
-	store.Delete(types.ConsumerIdToStopTimeKey(consumerId))
-}
-
-// getConsumerIdsBasedOnTime returns all the consumer ids stored under this specific `key(time)`
-func (k Keeper) getConsumerIdsBasedOnTime(ctx sdk.Context, key func(time.Time) []byte, time time.Time) (types.ConsumerIds, error) {
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(key(time))
-	if bz == nil {
-		return types.ConsumerIds{}, nil
-	}
-
-	var consumerIds types.ConsumerIds
-
-	if err := consumerIds.Unmarshal(bz); err != nil {
-		return types.ConsumerIds{}, fmt.Errorf("failed to unmarshal consumer ids: %w", err)
-	}
-	return consumerIds, nil
-}
-
-// appendConsumerIdOnTime appends the consumer id on all the other consumer ids under `key(time)`
-func (k Keeper) appendConsumerIdOnTime(ctx sdk.Context, consumerId string, key func(time.Time) []byte, time time.Time) error {
-	store := ctx.KVStore(k.storeKey)
-
-	consumers, err := k.getConsumerIdsBasedOnTime(ctx, key, time)
-	if err != nil {
-		return err
-	}
-
-	consumersWithAppend := types.ConsumerIds{
-		Ids: append(consumers.Ids, consumerId),
-	}
-
-	bz, err := consumersWithAppend.Marshal()
-	if err != nil {
-		return err
-	}
-
-	store.Set(key(time), bz)
-	return nil
-}
-
-// removeConsumerIdFromTime removes consumer id stored under `key(time)`
-func (k Keeper) removeConsumerIdFromTime(ctx sdk.Context, consumerId string, key func(time.Time) []byte, time time.Time) error {
-	store := ctx.KVStore(k.storeKey)
-
-	consumers, err := k.getConsumerIdsBasedOnTime(ctx, key, time)
-	if err != nil {
-		return err
-	}
-
-	if len(consumers.Ids) == 0 {
-		return fmt.Errorf("no consumer ids found for this time: %s", time.String())
-	}
-
-	// find the index of the consumer we want to remove
-	index := -1
-	for i := 0; i < len(consumers.Ids); i = i + 1 {
-		if consumers.Ids[i] == consumerId {
-			index = i
-			break
-		}
-	}
-
-	if index == -1 {
-		return fmt.Errorf("failed to find consumer id (%s)", consumerId)
-	}
-
-	if len(consumers.Ids) == 1 {
-		store.Delete(key(time))
-		return nil
-	}
-
-	consumersWithRemoval := types.ConsumerIds{
-		Ids: append(consumers.Ids[:index], consumers.Ids[index+1:]...),
-	}
-
-	bz, err := consumersWithRemoval.Marshal()
-	if err != nil {
-		return err
-	}
-
-	store.Set(key(time), bz)
-	return nil
-}
-
-// GetConsumersToBeLaunched returns all the consumer ids of chains stored under this spawn time
-func (k Keeper) GetConsumersToBeLaunched(ctx sdk.Context, spawnTime time.Time) (types.ConsumerIds, error) {
-	return k.getConsumerIdsBasedOnTime(ctx, types.SpawnTimeToConsumerIdsKey, spawnTime)
-}
-
-// AppendConsumerToBeLaunched appends the provider consumer id for the given spawn time
-func (k Keeper) AppendConsumerToBeLaunched(ctx sdk.Context, consumerId string, spawnTime time.Time) error {
-	return k.appendConsumerIdOnTime(ctx, consumerId, types.SpawnTimeToConsumerIdsKey, spawnTime)
-}
-
-// RemoveConsumerToBeLaunched removes consumer id from if stored for this specific spawn time
-func (k Keeper) RemoveConsumerToBeLaunched(ctx sdk.Context, consumerId string, spawnTime time.Time) error {
-	return k.removeConsumerIdFromTime(ctx, consumerId, types.SpawnTimeToConsumerIdsKey, spawnTime)
-}
-
-// GetConsumersToBeStopped returns all the consumer ids of chains stored under this stop time
-func (k Keeper) GetConsumersToBeStopped(ctx sdk.Context, stopTime time.Time) (types.ConsumerIds, error) {
-	return k.getConsumerIdsBasedOnTime(ctx, types.StopTimeToConsumerIdsKey, stopTime)
-}
-
-// AppendConsumerToBeStopped appends the provider consumer id for the given stop time
-func (k Keeper) AppendConsumerToBeStopped(ctx sdk.Context, consumerId string, stopTime time.Time) error {
-	return k.appendConsumerIdOnTime(ctx, consumerId, types.StopTimeToConsumerIdsKey, stopTime)
-}
-
-// RemoveConsumerToBeStopped removes consumer id from if stored for this specific stop time
-func (k Keeper) RemoveConsumerToBeStopped(ctx sdk.Context, consumerId string, stopTime time.Time) error {
-	return k.removeConsumerIdFromTime(ctx, consumerId, types.StopTimeToConsumerIdsKey, stopTime)
-}
-
 // GetOptedInConsumerIds returns all the consumer ids where the given validator is opted in
 func (k Keeper) GetOptedInConsumerIds(ctx sdk.Context, providerAddr types.ProviderConsAddress) (types.ConsumerIds, error) {
 	store := ctx.KVStore(k.storeKey)
@@ -434,111 +289,6 @@ func (k Keeper) RemoveOptedInConsumerId(ctx sdk.Context, providerAddr types.Prov
 	}
 
 	store.Set(types.ProviderConsAddrToOptedInConsumerIdsKey(providerAddr), bz)
-	return nil
-}
-
-// GetConsumersReadyToLaunch returns the consumer ids of the pending initialized consumer chains
-// that are ready to launch,  i.e., consumer clients to be created.
-func (k Keeper) GetConsumersReadyToLaunch(ctx sdk.Context, limit uint32) []string {
-	store := ctx.KVStore(k.storeKey)
-
-	spawnTimeToConsumerIdsKeyPrefix := types.SpawnTimeToConsumerIdsKeyPrefix()
-	iterator := storetypes.KVStorePrefixIterator(store, []byte{spawnTimeToConsumerIdsKeyPrefix})
-	defer iterator.Close()
-
-	result := []string{}
-	for ; iterator.Valid(); iterator.Next() {
-		spawnTime, err := types.ParseTime(spawnTimeToConsumerIdsKeyPrefix, iterator.Key())
-		if err != nil {
-			k.Logger(ctx).Error("failed to parse spawn time",
-				"error", err)
-			continue
-		}
-		if spawnTime.After(ctx.BlockTime()) {
-			return result
-		}
-
-		// if current block time is equal to or after spawnTime, and the chain is initialized, we can launch the chain
-		consumerIds, err := k.GetConsumersToBeLaunched(ctx, spawnTime)
-		if err != nil {
-			k.Logger(ctx).Error("failed to retrieve consumers to launch",
-				"spawn time", spawnTime,
-				"error", err)
-			continue
-		}
-		if len(result)+len(consumerIds.Ids) >= int(limit) {
-			remainingConsumerIds := len(result) + len(consumerIds.Ids) - int(limit)
-			if len(consumerIds.Ids[:len(consumerIds.Ids)-remainingConsumerIds]) == 0 {
-				return result
-			}
-			return append(result, consumerIds.Ids[:len(consumerIds.Ids)-remainingConsumerIds]...)
-		} else {
-			result = append(result, consumerIds.Ids...)
-		}
-	}
-
-	return result
-}
-
-// GetConsumersReadyToStop returns the consumer ids of the pending launched consumer chains
-// that are ready to stop
-func (k Keeper) GetConsumersReadyToStop(ctx sdk.Context, limit uint32) []string {
-	store := ctx.KVStore(k.storeKey)
-
-	stopTimeToConsumerIdsKeyPrefix := types.StopTimeToConsumerIdsKeyPrefix()
-	iterator := storetypes.KVStorePrefixIterator(store, []byte{stopTimeToConsumerIdsKeyPrefix})
-	defer iterator.Close()
-
-	result := []string{}
-	for ; iterator.Valid(); iterator.Next() {
-		stopTime, err := types.ParseTime(stopTimeToConsumerIdsKeyPrefix, iterator.Key())
-		if err != nil {
-			k.Logger(ctx).Error("failed to parse stop time",
-				"error", err)
-			continue
-		}
-		if stopTime.After(ctx.BlockTime()) {
-			return result
-		}
-
-		consumers, err := k.GetConsumersToBeStopped(ctx, stopTime)
-		if err != nil {
-			k.Logger(ctx).Error("failed to retrieve consumers to stop",
-				"stop time", stopTime,
-				"error", err)
-			continue
-		}
-		if len(result)+len(consumers.Ids) >= int(limit) {
-			remainingConsumerIds := len(result) + len(consumers.Ids) - int(limit)
-			if len(consumers.Ids[:len(consumers.Ids)-remainingConsumerIds]) == 0 {
-				return result
-			}
-			return append(result, consumers.Ids[:len(consumers.Ids)-remainingConsumerIds]...)
-		} else {
-			result = append(result, consumers.Ids...)
-		}
-	}
-
-	return result
-}
-
-// LaunchConsumer launches the chain with the provided consumer id by creating the consumer client and the respective
-// consumer genesis file
-func (k Keeper) LaunchConsumer(ctx sdk.Context, consumerId string) error {
-	err := k.CreateConsumerClient(ctx, consumerId)
-	if err != nil {
-		return err
-	}
-
-	consumerGenesis, found := k.GetConsumerGenesis(ctx, consumerId)
-	if !found {
-		return errorsmod.Wrapf(types.ErrNoConsumerGenesis, "consumer genesis could not be found for consumer id: %s", consumerId)
-	}
-
-	if len(consumerGenesis.Provider.InitialValSet) == 0 {
-		return errorsmod.Wrapf(types.ErrInvalidConsumerGenesis, "consumer genesis initial validator set is empty - no validators opted in consumer id: %s", consumerId)
-	}
-
 	return nil
 }
