+++ x/ccv/provider/keeper/permissionless_test.go
@@ -244,157 +244,6 @@ func TestConsumerPhase(t *testing.T) {
 	require.Equal(t, providertypes.ConsumerPhase_CONSUMER_PHASE_LAUNCHED, phase)
 }
 
-// TestConsumerStopTime tests the getter, setter, and deletion of the consumer id to stop times methods
-func TestConsumerStopTime(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	_, err := providerKeeper.GetConsumerStopTime(ctx, "consumerId")
-	require.Error(t, err)
-
-	expectedStopTime := time.Unix(1234, 56789)
-	providerKeeper.SetConsumerStopTime(ctx, "consumerId", expectedStopTime)
-	actualStopTime, err := providerKeeper.GetConsumerStopTime(ctx, "consumerId")
-	require.NoError(t, err)
-	require.Equal(t, actualStopTime, expectedStopTime)
-
-	providerKeeper.DeleteConsumerStopTime(ctx, "consumerId")
-	_, err = providerKeeper.GetConsumerStopTime(ctx, "consumerId")
-	require.Error(t, err)
-}
-
-// TestConsumersToBeLaunched tests `AppendConsumerToBeLaunched`, `GetConsumersToBeLaunched`, and `RemoveConsumerToBeLaunched`
-func TestConsumersToBeLaunched(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	spawnTime := time.Now()
-	err := providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId1", spawnTime)
-	require.NoError(t, err)
-	consumers, err := providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1"}, consumers.Ids)
-
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId2", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, consumers.Ids)
-
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId3", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, consumers.Ids)
-
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId2", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// also add consumer ids under a different spawn time and verify everything under the original spawn time is still there
-	spawnTimePlusOneHour := spawnTime.Add(time.Hour)
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId4", spawnTimePlusOneHour)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTimePlusOneHour)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId4"}, consumers.Ids)
-
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// start removing all consumers from `spawnTime`
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId3", spawnTime)
-	require.NoError(t, err)
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId1", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// remove from `spawnTimePlusOneHour`
-	err = providerKeeper.RemoveConsumerToBeLaunched(ctx, "consumerId4", spawnTimePlusOneHour)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTimePlusOneHour)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// add another consumer for `spawnTime`
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId5", spawnTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeLaunched(ctx, spawnTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId5"}, consumers.Ids)
-}
-
-// TestConsumersToBeStopped tests `AppendConsumerToBeLaunched`, `GetConsumersToBeLaunched`, and `RemoveConsumerToBeLaunched`
-func TestConsumersToBeStopped(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	stopTime := time.Now()
-	err := providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId1", stopTime)
-	require.NoError(t, err)
-	consumers, err := providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1"}, consumers.Ids)
-
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId2", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, consumers.Ids)
-
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId3", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, consumers.Ids)
-
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId2", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// also add consumer ids under a different stop time and verify everything under the original stop time is still there
-	stopTimePlusOneHour := stopTime.Add(time.Hour)
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId4", stopTimePlusOneHour)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTimePlusOneHour)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId4"}, consumers.Ids)
-
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId1", "consumerId3"}, consumers.Ids)
-
-	// start removing all consumers from `stopTime`
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId3", stopTime)
-	require.NoError(t, err)
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId1", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// remove from `stopTimePlusOneHour`
-	err = providerKeeper.RemoveConsumerToBeStopped(ctx, "consumerId4", stopTimePlusOneHour)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTimePlusOneHour)
-	require.NoError(t, err)
-	require.Empty(t, consumers.Ids)
-
-	// add another consumer for `stopTime`
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId5", stopTime)
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetConsumersToBeStopped(ctx, stopTime)
-	require.NoError(t, err)
-	require.Equal(t, []string{"consumerId5"}, consumers.Ids)
-}
-
 // TestOptedInConsumerIds tests the `GetOptedInConsumerIds`, `AppendOptedInConsumerId`, and `RemoveOptedInConsumerId` methods
 func TestGetOptedInConsumerIds(t *testing.T) {
 	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
@@ -447,74 +296,6 @@ func TestGetOptedInConsumerIds(t *testing.T) {
 	consumers, err = providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
 	require.NoError(t, err)
 	require.Empty(t, consumers)
-}
-
-// TestGetConsumersReadyToLaunch tests that the ready to-be-launched consumer chains are returned
-func TestGetConsumersReadyToLaunch(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	// no chains to-be-launched exist
-	require.Empty(t, providerKeeper.GetConsumersReadyToLaunch(ctx, 5))
-
-	err := providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId1", time.Unix(10, 0))
-	require.NoError(t, err)
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId2", time.Unix(20, 0))
-	require.NoError(t, err)
-	err = providerKeeper.AppendConsumerToBeLaunched(ctx, "consumerId3", time.Unix(30, 0))
-	require.NoError(t, err)
-
-	// time has not yet reached the spawn time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(9, 999999999))
-	require.Empty(t, providerKeeper.GetConsumersReadyToLaunch(ctx, 3))
-
-	// time has reached the spawn time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(10, 0))
-	require.Equal(t, []string{"consumerId1"}, providerKeeper.GetConsumersReadyToLaunch(ctx, 3))
-
-	// time has reached the spawn time of "consumerId1" and "consumerId2"
-	ctx = ctx.WithBlockTime(time.Unix(20, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, providerKeeper.GetConsumersReadyToLaunch(ctx, 3))
-
-	// time has reached the spawn time of all chains
-	ctx = ctx.WithBlockTime(time.Unix(30, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, providerKeeper.GetConsumersReadyToLaunch(ctx, 3))
-
-	// limit the number of returned consumer chains
-	require.Equal(t, []string{}, providerKeeper.GetConsumersReadyToLaunch(ctx, 0))
-	require.Equal(t, []string{"consumerId1"}, providerKeeper.GetConsumersReadyToLaunch(ctx, 1))
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, providerKeeper.GetConsumersReadyToLaunch(ctx, 2))
-}
-
-func TestGetConsumersReadyToStop(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	// no chains to-be-stopped exist
-	require.Empty(t, providerKeeper.GetConsumersReadyToStop(ctx, 3))
-
-	err := providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId1", time.Unix(10, 0))
-	require.NoError(t, err)
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId2", time.Unix(20, 0))
-	require.NoError(t, err)
-	err = providerKeeper.AppendConsumerToBeStopped(ctx, "consumerId3", time.Unix(30, 0))
-	require.NoError(t, err)
-
-	// time has not yet reached the stop time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(9, 999999999))
-	require.Empty(t, providerKeeper.GetConsumersReadyToStop(ctx, 3))
-
-	// time has reached the stop time of "consumerId1"
-	ctx = ctx.WithBlockTime(time.Unix(10, 0))
-	require.Equal(t, []string{"consumerId1"}, providerKeeper.GetConsumersReadyToStop(ctx, 3))
-
-	// time has reached the stop time of "consumerId1" and "consumerId2"
-	ctx = ctx.WithBlockTime(time.Unix(20, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2"}, providerKeeper.GetConsumersReadyToStop(ctx, 3))
-
-	// time has reached the stop time of all chains
-	ctx = ctx.WithBlockTime(time.Unix(30, 0))
-	require.Equal(t, []string{"consumerId1", "consumerId2", "consumerId3"}, providerKeeper.GetConsumersReadyToStop(ctx, 3))
 }
 
 func TestUpdateAllowlist(t *testing.T) {
